+++ app/models/agents/scheduler_agent.rb
@@ -87,10 +87,6 @@ module Agents
       true
     end
 
-    def check!
-      control!
-    end
-
     def validate_options
       if (spec = options['schedule']).present?
         begin