+++ spec/controllers/agents_controller_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe AgentsController do
   def valid_attributes(options = {})
@@ -397,6 +397,8 @@ describe AgentsController do
     it "accepts an event" do
       sign_in users(:bob)
       agent = agents(:bob_website_agent)
+      agent.options['url_from_event'] = '{{ url }}'
+      agent.save!
       url_from_event = "http://xkcd.com/?from_event=1".freeze
       expect {
         post :dry_run, id: agent, event: { url: url_from_event }