+++ spec/models/agents/peak_detector_agent_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe Agents::PeakDetectorAgent do
   before do