+++ spec/models/agents/sentiment_agent_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe Agents::SentimentAgent do
     before do